requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
motor==3.6.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
openai==1.52.2
